requirements.txt
langchain
wikipedia
langchain-community
langchain-anthropic
python-dotenv
pydantic
google-generativeai
requests
imapclient
mail-parser
beautifulsoup4
langchain-google-genai
transformers
tiktoken
langgraph
torch